requirements.txt
openai-whisper
translate-python
gtts
